Based the FPS counter's displayed value on the longest frametime
The previous behavior caused the counter to often display a very high FPS value due to an occasional long frame causing the delay between following frames to be very short. This commit makes the FPS counter more useful and intuitive for benchmarking purposes by calculating the displayed value based on the longest of the sampled frametimes.


Former-commit-id: 7650c897a2521d39349219a6fe0d0f486be0e80d
Former-commit-id: 5ab988623b97f291f0f7addeab846d3c8d3a246b
Former-commit-id: b38b30d717aecdcec6c6b6c18db52c3a7e1292e0
Former-commit-id: 091eeb4454e9089d21e43b9eb1f25c0c401d5b5d [formerly cb3d890815cc83aed13b5e22a994df8bdf691f79]
Former-commit-id: 8e8b74f5b9ec3f7b30f89acc5099ab5cadb64d39