Allow specifying the random shift range for terrain animation times
Previously, terrain graphics [image] random_start= only accepted a boolean value, but this commit makes it optionally also accept an integer value, which will be used as the range of random shift (in milliseconds) that the animation is allowed. This can be used for example to allow animations such as water to play only slightly out of sync.


Former-commit-id: 95ead89f8f460057d43ac51f60441375ca5a4b8b
Former-commit-id: a8b8ba4074581046e752b424116b5db7ce5e5805
Former-commit-id: 3ed5eedfc9a8ab238d95810f054459627b56b49c
Former-commit-id: 1f9c09d37eac60cfb145337425a14e048d120009 [formerly 50e03768e7825b85e483b9303edb6278bea9f705]
Former-commit-id: c562f21c148c85b693414a0b767fb56b97ab3bf3