Revert "Refactor out wesnothd_connection_ptr wrapper class"
This reverts commit 084c60d3c54dcd1b5d71d09e9aca9601722c6825 [formerly dc8207771a05221267a35b5be613a1807b05aafe] [formerly 9dcb1cf58149087acbd7196809d2e5b8d83eb839 [formerly a9fcb034218f6969e47f4046c390c009932bf58e]] [formerly 56d8422c1efe27f58c41b87b195c145dff5d43e1] [formerly 231e1180823670f247591f1269eb7db80f4e40e6] [formerly fb42dd8d708e504faa1d006dc648ed548b0533e7] and fixes #1928. It turns out this made it so
the connection was never closed since it was the wrapper object being destroyed that called the stop()
function.

Since a worker thread needed to hold a shared_ptr to the main ptr in order to ensure it wasn't destroyed
early, we can't just have a shared_ptr of wesnothd_connection directly, since the worker ref means it's
never destroyed.


Former-commit-id: 3b53ead45b29273e56e9c9cc876f06c192270915
Former-commit-id: 67e808849cf4db339d1e6adabf1140a452182786
Former-commit-id: 2715d9187179f43a6db01ac66965791f9dbd8816
Former-commit-id: 3ad812230a181de2a0859a561df2e5878ee79197 [formerly 6d5f02b6ebda934f70316e4375224d0b2a2a2c8c]
Former-commit-id: 9e5945fccab8981d01159a4f6a77eee601051c0a