Fixup 19bbfed again (fixs bug #1941)
TO quote @jyrkive: "prev_def is a pointer to the defendant after the previous fight. The defendant belongs
to bc, it's assigned here. And now when bc is a smart pointer, it frees the battle_context when it goes out
of scope, i.e. when the loop iteration changes. In the next iteration, the stale prev_def is accessed before
a new battle_context is created: use-after-free."


Former-commit-id: 171d161c7b2a2b37a40484c82c8a8f0232955e64
Former-commit-id: d43a9871e012e463c779ddf0f63862bc9c60869e
Former-commit-id: 00397c5cc88ee39884e73402fbb34644cfd9aa6e
Former-commit-id: 989302654e2a0a5319432fabdccb345d11348962 [formerly e6b93bcc5c44e8430daf661e47e0045b913e22f4]
Former-commit-id: 3f613348b801624f912982d2b8c462a903a99300