Fix freeze when launching web browser on Linux/macOS
POSIX has a feature to automatically reap zombie processes, which makes it
unnecessary to wait for the child process to exit. Activate the feature
and stop waiting for exit.

"POSIX.1-2001 specifies that if the disposition of SIGCHLD is set to
SIG_IGN or the SA_NOCLDWAIT flag is set for SIGCHLD (see sigaction(2)),
then children that terminate do not become zombies"


Former-commit-id: 1741ede147942d7e8134edb05a8ca0e9685d80a6
Former-commit-id: 55c49274c1deda5924682a05f63580cb62ac0ce7
Former-commit-id: 88fd404bab09556a1ca43a0c03c3197fe84d2f9b
Former-commit-id: 3d3e51b2b2739280fe91b270ed6a766d1a926b0d [formerly d7098c457c9bc334c3310bbdfcd6b432bcbcde79]
Former-commit-id: db182299d93a518749858c978ff7dbf33674df67