Refactor recording of completed difficulties
This stores the data in a completed_at key in the respective [difficulty] subtag instead of as a dedicated key
in the campaign config. This makes the code in Campaign Selection a bit easier to understand and will allow me
to implement the different colored laurels in the Campaign Difficulty dialog more easily.


Former-commit-id: 9b970b10813802ba55360c1f3660098915e8423f
Former-commit-id: 70e622b864dd30b250a877735ae3af67c021453d
Former-commit-id: 332523fbdcf27dabb3179c8c8a740156a19ae012
Former-commit-id: 04210da9c0a452959003ef144f2bd25364ca7d73 [formerly a168fb881f169ba7355c9589af6ff09b131ec88a]
Former-commit-id: 32cde50031c36cdad1ae9045e499c946d9a74b6f