Fix #1563: graphical glitch when two halos touch
I had added a unrender() call to halo_impl::effect::set_location() in an
attempt to fix haloes from showing in the top-left corner when a Mage of
Light is being moved. It had turned out to be unnecessary in the end,
but I had retained the call because I thought it was the right thing to do.

However, it turned out that removing that call is all it takes to fix bug
#1563.

Thanks to @newfrenchy83 for bringing the fix to my attention.


Former-commit-id: 92848ca61c4c1880c6e7ac948e4cb9a344827a4b
Former-commit-id: ceabdd5cadd2c109be65f4ad13ab6014f174b97f
Former-commit-id: 551f226adfd96c11361c915b9a8e1a23a0c0746f
Former-commit-id: a46161eb41a462b40181dcf401062bf2656f73e5 [formerly c11820ef1730a8f68622cdbe20f7eb399bb9c425]
Former-commit-id: 3556249bede216f7f032a8f9aa71f9c3270c7386