Fix #1987: halo remained after killing an unit with the context menu
The cause for this bug was the whiteboard.

A unit destroys its halo when the unit itself is destroyed. Makes sense,
right? Unfortunately, units are managed by shared pointers, and therefore
are not destroyed until all references to them are gone. And the whiteboard
retains a reference to the most recently selected unit.

The obvious fix would be to use a weak (non-owning) pointer in the
whiteboard, but it's sadly not possible because unit_ptr is an
intrusive_ptr. It stores the reference count right in the unit itself, and
therefore the refcount is lost when the unit is destroyed: hence weak
pointers are impossible.

Thus, what I ended up doing was explicitly notifying the whiteboard when an
unit is killed, and releasing the reference when it happens. If anyone has
suggestions about a better implementation, I'd love to hear them.


Former-commit-id: 9b10194e70bdfc4faf63332ee79e045e970e3262
Former-commit-id: 6c9d0f906f3fe088554c621c77a1e2abd970c2b7
Former-commit-id: e3b326997da96057516ca238277a7df716286fb0
Former-commit-id: f15821098322106f0263961563daced74213974b [formerly e1d5dc608abe136e63cd2c0c848ad4991dabbcc1]
Former-commit-id: 4518fe65d797c557b4b30689f38a1697732b5b33