campaignd: Send uploading clients a list of illegal names when any are found
(See issue #2043.)

It wouldn't be proper to stuff the full list into the conventional error
response, so in order to provide clients with the means to report it in
a better fashion in the future (a la WML load errors report dialog, with
a fancy scrolling box and an option to copy the report to clipboard and
stuff), we're attaching it as an extra_data attribute in the [error]
response. Naturally, only clients that are aware of its existence can
handle it, so this whole patch is completely unsuitable for 1.12.

Still, some level of backwards compatibility is retained -- the [error]
continues to be perfectly functional for incompatible clients (e.g.
1.13.10), and the details are simply missing for them.

The list would currently look like this:

  Test_Addon_420/~illegalname
  Test_Addon_420/testdir/illegal name 2
  Test_Addon_420/testdir/illegal~dir~name/

The required client-side changes for handling the extra_data attribute
and displaying its contents are included in this patch, but the GUI
changes are very rudimentary and should be considered more of a
proof-of-concept than a final solution, which I'll leave to someone who
actually groks GUI2 as it is right now and is an active developer in
the first place, which I am not.

Also note that the only reason the second parameter for
check_names_legal() is optional is that the function is also used
client-side by addons_client::install_addon(), which really doesn't need
to waste time building a full report of invalid names -- add-ons
containing those aren't supposed to make it to the server in the first
place, and the client-side check only exists as an extra safeguard.


Former-commit-id: fc20e4f3e2887dd0b73a87cc0cee77f0315c3921
Former-commit-id: c73e53b68e8c3eeafd71957b86fab66abbd2b109
Former-commit-id: 6a1d0fc7254b07bded12c605a8af8a19caeb092f
Former-commit-id: c461f761c95dfe0e72d9539ed26a420e1698ea04 [formerly 30b9a1c528c202e13b4365b5a23d71dbc5f34b86]
Former-commit-id: f3a96d620efb5c4cf2ce01e0867ea996a44c7dd3