AIs: idle_ai is not a "Dev AI", fixes #1955
As indicated in the comment of the idle_ai config file, an AI needs to
contain at least one stage, otherwise the default AI will be injected
under certain circumstances. Unlike indicated in the comment, this now
also applies to loading a mid-scenario save. (I will add that to the
comment in the next commit.)

As this AI was in data/ai/dev/, it was previously only loaded when
Wesnoth was started in debug mode. The config was thus not available
when playing in no-debug mode, causing the default AI to be injected on
reload. This caused issue #1955.

While not strictly necessary, we keep a redirect in data/ai/dev for
backward compatibility, in case the file is included in add-ons.


Former-commit-id: 7b03b0c166333aafa291da238c90410c3703d1f8
Former-commit-id: 95fb63a031880c24b94d90e4d46d078692c02fe7
Former-commit-id: 72d078401e2839925ed1444de97bce4fdc79e690
Former-commit-id: 03fcee26b3d9db20fe20de1733d69cef8a310b06 [formerly 169c97f3ee655f925bf9289dcd6fc11cc13b0506]
Former-commit-id: 8f6980e110b76aea1704da9bba268688752433e1