DM S09: Fix issues by rebalancing scenario
Give player control of the ghosts and ghouls as they are found.
Update dialog to match ghosts and ghouls joining Delfador.
Enable gold carryover so player doesn't pointlessly wait till last turn.
Lower turn limit, as board is easier with additional units on player side.
Set start gold to 0, so final carryover is lower. With gold and turn limit
changes, it lessens amount in next scenario, but 300 is still achievable,
which is plenty.

Add cave wall to map around tomb. Otherwise a ghost going through the
secret passage causes an akward triggering of the village warning dialog.
Also makes it harder to get to the staff without using the secret passage.


Former-commit-id: 8fad8842daf2fbc3861aecc8bf963ada95da44e5
Former-commit-id: 2c7fc7489adc2b8c4664489a24c4dcca0414e596
Former-commit-id: 6b1a32b301f304835a53838aebad7ec802572736
Former-commit-id: d5795938b96cba8f916ffe43a2d6549450d00691 [formerly 4ec7381b3688142fc237172e9323da65940f2364]
Former-commit-id: 9631669b7ac2475b3721b4b4f74f10ad5aaae4fe