Reconfigure 'select all' for texts in macOS.
Use the idiomatic combination.

Update reference to particular comment line of
'src/widgets/textbox.cpp' in comment with the current line where the
comment is now.

Remove `unsigned` as the type for `copypaste_modifier`, because it was
a little bit less future proof that way, it is better as an
`SDL_Keycode`, which at the moment is `Sint32`, which in turn is
`int32_t` ultimately. Reuse the constant for 'select all' so renaming
it to `modifier_key`.


Former-commit-id: feda4c462b4918a11e47b1a0d521f3103774e6e4
Former-commit-id: ee593d376ecf7c8a76f66356fa3950ff5f86ecd5
Former-commit-id: 1b88894ebc5b0bdcb2f61f0ec1b5958881bb500d
Former-commit-id: 0ec38e9b5dedf748ed379360877925b7dcf19494 [formerly b762773824977496792a49b2fe86d9480d9e99dd]
Former-commit-id: 164d0dc60fc0c2ad59cb1c7e761a1660b4aeaf98