Remove Unit Box and Widescreen themes
[ci skip]

These were half-baked and mostly badly-designed. The Unit Box theme was never finished, is completely broken,
and if we were to do a major theme refresh we'd probably go in another direction. The Widescreen theme was a
legacy experiment that was never adopted and as of now is basically alsp completely broken.


Former-commit-id: 7bd99777e5543ca4505ac69fd51fcefd21e187c0
Former-commit-id: 55bfd7019f927622242049058f32785bf4be02cc
Former-commit-id: aa3c93d503df694f6481b193e1b365b3f1aa3967
Former-commit-id: 96e45d8c3f5beb0fe17a9e0f797297cf8f722e31 [formerly e46ae073c0c46a8029278b3453c9b8a4003afd85]
Former-commit-id: 17f49fcf07aee703dbfcb0caaff32fdc23694ced