Travis CI: globally disable C++ unit tests
The tests haven't been run for a couple of months.
The test executable crashes on startup with the error message

Xlib: extension "RANDR" missing on display ":99.0".

After ten attempts, the test wrapper just gives up and returns success.

So, let's not even pretend that we have C++ unit tests running in Travis.
We just don't. Explicitly disabling the tests makes it more clear.


Former-commit-id: bccce97b16d698ddfed8b567263e04ca2efd7b20
Former-commit-id: 503b26bbf9a44f782ef9c189c285c001d5d1bf1d
Former-commit-id: 84f867dcdf3a6eada63cb6874cb044368fe00775
Former-commit-id: 7b3d2097d136c7676860be615db1b139e6146427 [formerly 493a9d95a0ec89fff008863211d9ab5e19a453eb]
Former-commit-id: 568bdba73f0f145d25ac677dd9e2499028c120c1