Connect Engine: removed "--give--" entry from controller options
This was basically used to separate players from the general controller options. In the GUI1 dialog it
was unselectable. I never got around to implementing that in the GUI2 dialog, but regardless, it doesn't
look great in the menu so I'm removing it. Plus, since the dropdowns are GUI2, we could apply different
text formatting to denote players if we wanted.


Former-commit-id: dff415242b0c1fef56042003025e1e13b038b4f9
Former-commit-id: 4c334eadfb305dddf4075bf4c9ec2bb5aa4c9ff9
Former-commit-id: 6d4aa8816360cc969e26f35f34d8784f8aec25b3
Former-commit-id: d49a9ea0819bcde66772055fc87e3c60594c9b95 [formerly d6988ee0e69a6b40fab7708282f31adb635a2e20]
Former-commit-id: a6714cfa16cb729880d52bde4039a1e05ce882bb