campaignd: Blacklist standard reserved DOS Device Names
Windows XP and later define DOS device names for several system devices
that will be handled as if they had directory entries everywhere. Their
names may not be used on the left side of a file name if a dot is
present, otherwise the read or write operation in progress (if
applicable) is redirected to the device in question, with unexpected
results (e.g. writing to CON writes to a console that may or may not be
attached, writing to COM1 will send data through the first enumerated
serial port, and so on).

Thus, we really don't want add-on authors to either intentionally or
accidentally use them on non-Windows platforms and cause Windows-based
clients to do unexpected things.

Emitting special messages for the UI in this case seems unnecessary and
it'll be a waste of space most of the time since these names should crop
up extremely rarely. Better just tell people what's up on a case-by-case
basis if they decide to post in the forums about it.


Former-commit-id: 9476d051bbf0cbf9ef692cd2c8ecdbc3717130d5
Former-commit-id: 62bfdcc639b5bf55404d19cc59089a8c018935e4
Former-commit-id: 5b4e84c109f0dfb62dad021bf45092aa4aeaad33
Former-commit-id: bcde1e8ab905850d0b7be81e558da3a58775c1cf [formerly 7f36ec0228a832b83bd667c7ab220ecccfba5cbc]
Former-commit-id: dc7e5c1384de9c0049001faf44e123da230deed9