Some refactoring of team color data handling
These changes cover areas from PR #2101, excluding the removal of the numeric ID color ranges.
That can't be done yet (it breaks things).

This also excludes any changes to the connect_engine color handling. That needs to be refactored
separately. Despite the fact that it's done very similar to the new get_side_color_id function,
I need to deal with the fact that every side_engine member in the connect engine keeps a copy of
the default color options.

* Added a general-purpose function to extract color data from a config.
* Refactored the team class's general get-side's-color-data function so it always returns a color
  ID instead of falling back on a number.
* Deployed this change for the leader image coloring in the save_index; said coloring is now always
  done by ID.
* Save color data by ID in team_info instead of using the side number if no color was provided.
* Renamed get_side_color_index to get_side_color_id to better reflect its purpose.


Former-commit-id: 1b3c94087c1b9ef4d29ef137aff84d0e33a13d50
Former-commit-id: f8b894ace6103a0e679e38eac36e85c14270a619
Former-commit-id: 718d0b15effdaba2905972adb11ac3e5fdc693ce
Former-commit-id: f955584e0137d9b4dd2cce1d2dfde99f6882d63e [formerly b249a4036f5c4acf86b748594f1e29a3a7ae9c6f]
Former-commit-id: f8a5eb23df3396babfd705df26245a40c409a1f0