Globally suppress -Wmaybe-uninitialized
We have needed to suppress it in a bunch of places because GCC has kept
giving false positives depending on optimization settings. When I noticed
yet another false positive when I attempted to build the game with PGO
profiling data using CMake, it was the last straw.

Let's just suppress the warning globally and move on with our lives.


Former-commit-id: 6d1ffaffe4c3cfdf47fd0cd0dd8c9d225fd4225a
Former-commit-id: 52a99cfe1747d5314435850cff880226e3fe59c5
Former-commit-id: 73ff8b63fdc35b56d584bdae0bd7587a7bdd8291
Former-commit-id: 541fcb4b50d4a47d06aad1e4445379ab947f1f13 [formerly a14ea0583d60863791503e2073c324ac30996128]
Former-commit-id: aceafcb1894d0c9ecfc3bfdf6bf017cc677fe6f0