Revert "Revert "Disallow specifying file paths with backslashes""
This reverts commit 3dc3a3e45af9227e3c59e3fec62060823272c16b [formerly 87839111a56c249652870474ea665e5a351a346a] [formerly a88e16cba6f1bd3dfe0e88de819a2ea4a461f675 [formerly 4f694be13d3748da55c869483cc76c875a65d342]] [formerly c99a0f66affb0df0b890ff53f86ac631db7c5d1a] [formerly f53f1c7a06eb5582d2c6cd3ebca0f10d2b1c0b28] [formerly f2c5aa1d2e55372888c582d6f8bc9abe34ce7095].

Mainline does *not* contain paths with backslashes. If it did, the game
would be broken on GNU/Linux and macOS (and we'd need to fix it anyway).
Instead, it's more likely that we have code somewhere that generates
paths dynamically using backslashes on Windows, and the correct fix is
to change that code to use forward slashes instead.


Former-commit-id: c6b70381768b0acee4782ce0dab0e0f39b67cf1a
Former-commit-id: d8a0200430c1f1c79027ba06c093fb4ab4d3e230
Former-commit-id: 13638e61c06b45fda7cb1b2ff8e683fdbb23157d
Former-commit-id: 70f3c6303128efae8792efbde20ac039add6c42a [formerly cc1662732be969dc11c20c5cd661e709d4b9f8b4]
Former-commit-id: a4b554871a13c8da5c9e69c3f752b52d3ba199be