Fix Scons for SDL2_mixer 2.0.2
With the change from SDL_mixer 2.0.1 to 2.0.2, we cannot simply check if SDL_mixer can open an OGG/Vorbis file, we must first open an audio device. [This is probably a mistake, since the documentation implies simply initializing the codec should be enough.] To avoid future problems, upgrade the feature-test program to perform a full initialization and shut-down; everything except actually playing the sound.

Closes #2137


Former-commit-id: 379bdc2752ab6010b52685fc13300c88472a0479
Former-commit-id: 1d51e277194e3c0a23e32633421c44bcbb5edf2c
Former-commit-id: e6629ae734ab31049a46057264ecbd4cc7ccd9ff
Former-commit-id: 3da7c9b907bc892937df85a60be26f10597f5bff [formerly 529d1a1ce73a58fbec8099cc21a25907bf17fa05]
Former-commit-id: 13b558b092f9acbb421d877f2d80ec204ad3d528