MP Lobby: right-align status text
[ci skip]

This prevents status text from sometimes floating off to the left if the status label shrinks but
a relayout hasn't occurred. In that case, the label widget's area would be larger than needed for
the text, and since left-align is the default setting, the next text gets stuck weirdly aligned.


Former-commit-id: 0e37d045e7013a5055db39b2588bc846febdf03c
Former-commit-id: fa8f41779fa8dd8e8ef7701222fda5d4bc56d800
Former-commit-id: 92b547a1da6e62cdf8fec155be3577f6553aaf7f
Former-commit-id: 8f8ce878df6e9d589c4a0e2d0c3b67a3aeb05fb6 [formerly 9d9632914310d9fa8be1c00a361d7f08bcf4a8d8]
Former-commit-id: 90b47b69e787514e77f6b346bc9b3eea5e05ecb7